### 2021-10-27 #3863 @whedon
[PRE REVIEW]: 3D Line Radiative Transfer & Synthetic Observations with Magritte
**Submitting author:** @FredDeCeuster (<a href="http://orcid.org/0000-0001-5887-8498">Frederik De Ceuster</a>)
**Repository:** <a href="https://github.com/Magritte-code/Magritte" target ="_blank">https://github.com/Magritte-code/Magritte</a>
**Version:** 0.0.1
