+++ dependencies/conda_env.yml
@@ -1,7 +1,7 @@
 name: magritte
 channels:
-  - defaults
   - conda-forge
+  - defaults
 dependencies:
   - jupyterlab
   - yt
