### 2021-10-28 071a048
Use miniconda with python 3.7 instead of latest.
