### 2021-10-29 3ceed5c
Update get_conda.sh
