+++ setup.py
@@ -68,5 +68,5 @@ setup(
         "Topic :: Scientific/Engineering :: Astronomy",
         "Topic :: Scientific/Engineering :: Physics"
     ],
-    python_requires='>=3.6',
+    python_requires='>=3.7',
 )
