+++ dependencies/conda_env_2.yml
@@ -1,0 +1,284 @@
+name: magritte
+channels:
+  - conda-forge
+  - defaults
+dependencies:
+  - _libgcc_mutex=0.1=conda_forge
+  - _openmp_mutex=4.5=1_gnu
+  - alabaster=0.7.12=py_0
+  - alsa-lib=1.2.3=h516909a_0
+  - anyio=3.3.4=py37h89c1867_0
+  - aom=3.2.0=h9c3ff4c_2
+  - argcomplete=1.12.3=pyhd8ed1ab_2
+  - argon2-cffi=21.1.0=py37h5e8e339_0
+  - astropy=4.3.1=py37hb1e94ed_1
+  - async_generator=1.10=py_0
+  - attrs=21.2.0=pyhd8ed1ab_0
+  - babel=2.9.1=pyh44b312d_0
+  - backcall=0.2.0=pyh9f0ad1d_0
+  - backports=1.0=py_2
+  - backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
+  - bleach=4.1.0=pyhd8ed1ab_0
+  - breathe=4.31.0=pyhd8ed1ab_0
+  - brotlipy=0.7.0=py37h5e8e339_1001
+  - bzip2=1.0.8=h7f98852_4
+  - c-ares=1.18.1=h7f98852_0
+  - ca-certificates=2021.10.8=ha878542_0
+  - cached-property=1.5.2=hd8ed1ab_1
+  - cached_property=1.5.2=pyha770c72_1
+  - certifi=2021.10.8=py37h89c1867_0
+  - cffi=1.14.6=py37h036bc23_1
+  - cfitsio=3.470=hb418390_7
+  - cftime=1.5.1=py37hb1e94ed_0
+  - chardet=4.0.0=py37h89c1867_1
+  - cmarkgfm=0.6.0=py37h5e8e339_0
+  - colorama=0.4.4=pyh9f0ad1d_0
+  - cryptography=35.0.0=py37hf1a17b8_1
+  - curl=7.79.1=h2574ce0_1
+  - cycler=0.11.0=pyhd8ed1ab_0
+  - dbus=1.13.6=h48d8840_2
+  - debugpy=1.4.1=py37hcd2ae1e_0
+  - decorator=5.1.0=pyhd8ed1ab_0
+  - defusedxml=0.7.1=pyhd8ed1ab_0
+  - docutils=0.17.1=py37h89c1867_0
+  - double-conversion=3.1.5=h9c3ff4c_2
+  - eigen=3.4.0=h4bd325d_0
+  - entrypoints=0.3=pyhd8ed1ab_1003
+  - expat=2.4.1=h9c3ff4c_0
+  - ffmpeg=4.4.0=h6987444_5
+  - fontconfig=2.13.1=hba837de_1005
+  - freetype=2.10.4=h0708190_1
+  - gettext=0.19.8.1=h73d1719_1008
+  - gitdb=4.0.9=pyhd8ed1ab_0
+  - gitpython=3.1.24=pyhd8ed1ab_0
+  - gl2ps=1.4.2=h0708190_0
+  - glew=2.1.0=h9c3ff4c_2
+  - glib=2.70.0=h780b84a_1
+  - glib-tools=2.70.0=h780b84a_1
+  - gmp=6.2.1=h58526e2_0
+  - gmpy2=2.1.0b5=py37h025e8b9_0
+  - gnutls=3.6.13=h85f3911_1
+  - gst-plugins-base=1.18.5=hf529b03_1
+  - gstreamer=1.18.5=h9f60fe5_1
[…]
