+++ dependencies/conda_env.yml
@@ -3,6 +3,7 @@ channels:
   - conda-forge
 dependencies:
   - python=3.7
+  - numpy=1.19
   - jupyterlab
   - yt
   - scipy
