+++ dependencies/conda_env.yml
@@ -1,12 +1,14 @@
 name: magritte
 channels:
   - conda-forge
+  - defaults
 dependencies:
   - python=3.7
+  - numpy=1.19
+  - vtk
   - jupyterlab
   - yt
   - scipy
-  - vtk
   - mpi4py
   - plotly
   - pyyaml
