+++ dependencies/conda_env.yml
@@ -4,7 +4,7 @@ channels:
   - defaults
 dependencies:
   - python=3.7
-  - numpy=1.19
+  - numpy
   - vtk
   - jupyterlab
   - yt
