+++ dependencies/conda_env.yml
@@ -2,6 +2,7 @@ name: magritte
 channels:
   - conda-forge
   - defaults
+# The order of the channels above is critical!!!
 dependencies:
   - python=3.7
   - numpy
