+++ docs/src/1_examples/2_synthetic_observations/index.rst
@@ -15,8 +15,9 @@ The following examples show how to create synthetic observations with Magritte.
    :maxdepth: 1
    :caption: Contents
    
-   0_image_1D_AMRVAC.ipynb
-   1_image_3D_AMRVAC.ipynb
-   2_image_3D_AMRVAC_red.ipynb
-   3_image_3D_Phantom.ipynb
-   4_image_3D_Phantom_red.ipynb
+   0_image_analytic_spiral.ipynb
+   1_image_1D_AMRVAC.ipynb
+   2_image_3D_AMRVAC.ipynb
+   3_image_3D_AMRVAC_red.ipynb
+   4_image_3D_Phantom.ipynb
+   5_image_3D_Phantom_red.ipynb
