### 2021-11-03 8d3406f
First commit for new multiresolution scheme 'only using' the original grid
