+++ src/bindings/pybindings.cpp
@@ -91,6 +91,7 @@ PYBIND11_MODULE (core, module)
         .def_readwrite ("geometry",       &Model::geometry)
         .def_readwrite ("chemistry",      &Model::chemistry)
         .def_readwrite ("lines",          &Model::lines)
+        .def_readwrite ("using_same_grid",&Model::using_same_grid)
         .def_readwrite ("a",          &Model::a)
         .def_readwrite ("b",          &Model::b)
         .def_readwrite ("c",          &Model::c)
