+++ src/model/model.hpp
@@ -133,7 +133,7 @@ struct Model
     {
         cout << "c.size() = " << c().vec.size() << endl;
 
-        accelerated_for (i, c().vec.size(), 
+        accelerated_for (i, c().vec.size(),
         {
             c()[i] = a()[i] + b()[i];
 
