### 2021-11-03 #3863 @xuanxu
@mdelorme, @gonsie would you be willing to review this paper? I think your knowledge of Simulation software and C++/Python makes both of you a good fit for this submission.
