### 2021-11-04 89e15da
Added disk example to docs.
