### 2021-11-04 cfaca24
Able to compile nanoflan with magritte (not yet doing anything)
