### 2021-11-05 27fe468
Added nanoflann for kd-tree computations. TODO: figure out what part is going wrong
