+++ src/model/model.cpp
@@ -583,13 +583,126 @@ int Model :: compute_level_populations_multiresolution (
 
             if (using_same_grid)//if using the same grid, we need to interpolate for the other non-calculated points
             {
-                Size current_lvl=mrControllerHelper.get_current_level();
-                std::cout<<"current level"<<current_lvl<<std::endl;
-                //now interpolate for all other levels
-                for (Size level_diff=0; level_diff<current_lvl; level_diff++)
+                interpolate_levelpops_local(mrControllerHelper.get_current_level(),0);
+            //     Size current_lvl=mrControllerHelper.get_current_level();
+            //     std::cout<<"current level"<<current_lvl<<std::endl;
+            //     //now interpolate for all other levels
+            //     for (Size level_diff=0; level_diff<current_lvl; level_diff++)
+            //     {
+            //         interpolate_levelpops_local(current_lvl-level_diff);
+            //     }
+            //
+            //     // Again, do not forget to set emissivity and opacity after interpolating the levelpops
+            //     lines.set_emissivity_and_opacity ();
+            //
+            }
+
+
+            //If enabled, we now write the level populations to the hdf5 file
+            if (parameters.writing_populations_to_disk){
+                IoPython io = IoPython ("hdf5", parameters.model_name());
+                lines.write_populations_of_iteration(io, iterations_per_level[geometry.points.multiscale.get_curr_coars_lvl()], geometry.points.multiscale.get_curr_coars_lvl());
+                std::cout<<"Wrote populations to disk"<<std::endl;
+            }
+
+            //And now we check for all species whether the level populations have already converged
+            for (int l = 0; l < parameters.nlspecs(); l++)
+            {
+                //note: this information is only based on the points currently in the grid.
+                error_mean.push_back (lines.lineProducingSpecies[l].relative_change_mean);
+                error_max .push_back (lines.lineProducingSpecies[l].relative_change_max);
+
+                if (lines.lineProducingSpecies[l].fraction_not_converged > 0.005)
                 {
-                    interpolate_levelpops_local(current_lvl-level_diff);
+                    some_not_converged = true;
                 }
+
+                const double fnc = lines.lineProducingSpecies[l].fraction_not_converged;
+                prev_it_frac_not_converged[l]=fnc;
+
+                cout << "Already " << 100 * (1.0 - fnc) << " % converged!" << endl;
+            }
+
+            //If all level populations have converged, let it know to the mrController
+            if (!some_not_converged)
+            {
+                iteration_sum+=iteration;
+                iteration=0;
+                //signal convergence on current grid
+                mrControllerHelper.converged_on_current_grid();
+                std::cout<<"Signaling convergence"<<std::endl;
+            }
+
+            break;
+        }
+
+
+        //Signal to do one iteration the current level and interpolate if using_same_grid
+        case MrController::Actions::stay_and_interpolate:
+        {
+            //Stopping current timer and replace with timer at current level
+            (*current_timer_pointer).stop();
+            timers_per_level[geometry.points.multiscale.get_curr_coars_lvl()].start();
+            current_timer_pointer=&timers_per_level[geometry.points.multiscale.get_curr_coars_lvl()];
+
+            std::cout<<"Action stay"<<std::endl;
+
+            iterations_per_level[geometry.points.multiscale.get_curr_coars_lvl()]++;
+            iteration++;
+            // logger.write ("Starting iteration ", iteration);
+            cout << "Starting iteration " << iterations_per_level[geometry.points.multiscale.get_curr_coars_lvl()] << " for coarsening level: " << geometry.points.multiscale.get_curr_coars_lvl() << endl;
+
+
[…]
