### 2021-11-08 587158a
It seems that the new multiresolution methods do not work that well (they do not calculate the true solution at any coarser level, so it might have been expected). Either way, this could still form the basis for our emulator approach.
