### 2021-11-09 2377f5a
First implementation of radiative cooling calculation
