### 2021-11-12 #3905 @whedon
[REVIEW]: 3D Line Radiative Transfer & Synthetic Observations with Magritte
**Submitting author:** @FredDeCeuster (<a href="http://orcid.org/0000-0001-5887-8498">Frederik De Ceuster</a>)
**Repository:** https://github.com/Magritte-code/Magritte
**Branch with paper.md** (empty if default branch): joss_paper_1.0
**Version:** 0.0.1
