### 2021-11-12 fa6ba1b
Corrected implementation of the collisional cooling rate.
