### 2021-11-24 8ca012d
First preparations for the collocation solver.
