### 2021-11-25 f811d88
More preparations for the collocation solver. TODO: reduce memory requirement
