+++ src/solver/collocation/collocation.hpp
@@ -2,6 +2,7 @@
 
 #include "model/model.hpp"
 #include "model/parameters/parameters.hpp"
+#include <set>
 
 // Helper class for setting up the collocation solver
 class Collocation
@@ -12,17 +13,33 @@ private:
     const Real TRUNCATION_SIGMA=5.0; //< if Delta(freq)>trunc_sigma*freq_width just set the result from the gaussian basis function to zero (as it is almost zero)
 
     //NOTE TO SELF: FREQUENCIES (ESPECIALLY WIDTH) MIGHT CHANGE TO ALSO BE ON THE POINT ID (DUE TO TEMPERATURE DIFFERENCES)
-    vector<Real> frequencies;
-    vector<Real> frequencies_widths;
+    vector<vector<vector<Real>>> frequencies;//For every direction, for each point, the ordered DOPPLER SHIFTED frequencies
+    vector<vector<vector<Real>>> frequencies_widths;
+
+    //TODO: this thing is of the size of the actual matrix; maybe just make an approximation with respect to the actual doppler shift compared to the neighbors...
+    // so simply calculate the max velocity difference, then the max doppler shift is bounded (multiplicatively) by ~1+||Delta v||/c
+    // the advantage will be that we no longer care about the exact direction, nor about the exact other location
+    vector<vector<vector<vector<std::set<Size>>>>> other_frequency_basis_interacting_with; //< contains for every direction, every point, for every interacting rbf with that point, for each frequency basis, the indices of the frequency basis functions interacting with it of the interacting point
+    // NOTE TO DEV: might change to be also dependent on the exact point (temperature differences)
+    // Also, this (and the same thing for the rbf) could be replaced by a vector<set<>>; as we do not particularly care for the order
 
     const Size N_POINTS_IN_RADIAL_BASIS=16;
 
     vector<Vector3D> point_locations;
+    // vector<Size> point_indices; //< for transforming back to the regular point indices used in model (as we might only need a reduced set of all the points)
     vector<Real> rbf_radii;
-    vector<vector<Size>> other_basis_interacting_with; //< contains for each rbf, the indices of the radial basis functions interacting with it (also includes itself)
+    vector<vector<Size>> other_radial_basis_interacting_with; //< contains for each rbf, the indices of the radial basis functions interacting with it (also includes itself)
     //useful for transforming from basis space to actual solution space
 
     //for the directions, the basis functions are 1, so no need to store anything
+
+
+    inline void set_interacting_bases(Model& model);
+
+    //returns whether the frequencies do not lie to far from eachother
+    inline bool freq_cutoff_condition(Real abs_freq_diff, Real freq_basis_width);
+    // //returns whether the points do not lie to far from eachother
+    // inline
 
 
 public:
@@ -34,11 +51,11 @@ public:
 // The complete radial basis function at an index (dir, nu, pid) is a product of the ones defined below
 // basis functions for dir, nu, pid(implicitly x,y,z)
     inline Real basis_direction(Size rayindex);
-    inline Real basis_freq(Size freqidx, Real currnu);
+    inline Real basis_freq(Size rayidx, Size freqidx, Size pointidx, Real currfreq);
     inline Real basis_point(Size pointid, Vector3D& location);
 
 // basis function derivatives for nu, position
-    inline Real basis_freq_der(Size freqidx, Real currfreq);
+    // inline Real basis_freq_der(Size freqidx, Real currfreq);
     inline Real basis_point_der(Size centerpoint, Vector3D& location, Size rayindex, Geometry& geometry);
 
 // somewhere to store the sparse matrix and the decomposition TODO: figure out what we actually want
[…]
