### 2021-11-26 fcaa727
Even more background stuff for the collocation solver.
