+++ src/solver/collocation/collocation.hpp
@@ -3,6 +3,7 @@
 #include "model/model.hpp"
 #include "model/parameters/parameters.hpp"
 #include <set>
+#include <map>
 
 // Helper class for setting up the collocation solver
 class Collocation
@@ -13,25 +14,34 @@ private:
     const Real TRUNCATION_SIGMA=5.0; //< if Delta(freq)>trunc_sigma*freq_width just set the result from the gaussian basis function to zero (as it is almost zero)
 
     //NOTE TO SELF: FREQUENCIES (ESPECIALLY WIDTH) MIGHT CHANGE TO ALSO BE ON THE POINT ID (DUE TO TEMPERATURE DIFFERENCES)
-    vector<vector<vector<Real>>> frequencies;//For every direction, for each point, the ordered DOPPLER SHIFTED frequencies
-    vector<vector<vector<Real>>> frequencies_widths;
+    vector<vector<vector<Real>>> frequencies;//For every direction, for each point, the DOPPLER SHIFTED frequencies of the line transitions
+    // THESE ARE NOT ORDERED
+    vector<vector<vector<Real>>> frequencies_inverse_widths; //the corresponding inverse widths of the line transition gaussians
+    // Dev note: it is defined as 1/(sqrt(2)*sigma) for use in the traditional gaussian
 
-    //TODO: this thing is of the size of the actual matrix; maybe just make an approximation with respect to the actual doppler shift compared to the neighbors...
-    // so simply calculate the max velocity difference, then the max doppler shift is bounded (multiplicatively) by ~1+||Delta v||/c
+    // //TODO: this thing is of the size of the actual matrix; maybe just make an approximation with respect to the actual doppler shift compared to the neighbors...
+    // // so simply calculate the max velocity difference, then the max doppler shift is bounded (multiplicatively) by ~1+||Delta v||/c
+    // // the advantage will be that we no longer care about the exact direction, nor about the exact other location
+    // vector<vector<vector<vector<std::set<Size>>>>> other_frequency_basis_interacting_with; //< contains for every direction, every point, for every interacting rbf with that point, for each frequency basis, the indices of the frequency basis functions interacting with it of the interacting point
+
+    // We make an approsimation by simply calculating the max velocity difference, then the max doppler shift is bounded (multiplicatively) by ~1+||Delta v||/c
     // the advantage will be that we no longer care about the exact direction, nor about the exact other location
-    vector<vector<vector<vector<std::set<Size>>>>> other_frequency_basis_interacting_with; //< contains for every direction, every point, for every interacting rbf with that point, for each frequency basis, the indices of the frequency basis functions interacting with it of the interacting point
+    vector<vector<std::set<Size>>> other_frequency_basis_interacting_with; //< contains for every point, for each frequency basis, the indices of the frequency basis functions interacting with it of the interacting point
     // NOTE TO DEV: might change to be also dependent on the exact point (temperature differences)
     // Also, this (and the same thing for the rbf) could be replaced by a vector<set<>>; as we do not particularly care for the order
 
     const Size N_POINTS_IN_RADIAL_BASIS=16;
 
     vector<Vector3D> point_locations;
-    // vector<Size> point_indices; //< for transforming back to the regular point indices used in model (as we might only need a reduced set of all the points)
+    vector<Size> index_conversion; //< for transforming back to the regular point indices used in model (as we might only need a reduced set of all the points)
+    std::map<Size,Size> reverse_index_conversion; //< for transforming from the regular point indices to the point indices used here
     vector<Real> rbf_radii;
[…]
