### 2021-11-27 #3905 @gonsie
@FredDeCeuster can you clarify the author list on the paper?

In the readme Atulit Srivastava is listed as a developer, but I can't find any contributions from them in GitHub.

In the readme Ward Homan and Jan Bolte are listed as technical advisors, but they are not included in the author list.
