### 2021-11-28 Magritte-code/Magritte#52 @gonsie
Contributing guide (joss review)
This issue is related to openjournals/joss-reviews#3905.

This repo is lacking a [contributing guide](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors)
