### 2021-11-29 #3905 @FredDeCeuster
@gonsie

Atulit Srivastava developed parts for a future version (currently not public), but also already contributed ideas for this version.

Ward Homan and Jan Bolte mostly contributed to the science done with Magritte (see papers in documentation), but I would be happy to include them as co-authors on this paper.
