### 2021-11-29 Magritte-code/Magritte#54 @CFGrote
Example notebook `0_creating_models.ipynb` not working (JOSS Review)
I installed Magritte according to documentation and then opened the above mentioned example notebook.

Line 14 returns
```
FileNotFoundError: [Errno 2] No such file or directory: 'Analytic_disk/analytic_disk_convert_to_pos.geo_unrolled'
/bin/sh: 1: gmsh: not found
```

I suspected this came up because gmsh is installed via conda, so Popen.subprocess not being aware of a conda env might raise. Indeed, installing gmsh systemwide gets me one step further.

Rerunning the same cell now produces:

```
Info    : Running 'gmsh -0 Analytic_disk/analytic_disk_convert_to_pos.geo' [Gmsh 4.1.3, 1 node, max. 1 thread]
Info    : Started on Mon Nov 29 14:05:54 2021 Info    : Reading 'Analytic_disk/analytic_disk_convert_to_pos.geo'...
Info    : Done reading 'Analytic_disk/analytic_disk_convert_to_pos.geo'
Info    : Writing 'Analytic_disk/analytic_disk_convert_to_pos.geo_unrolled'...
Info    : Done writing 'Analytic_disk/analytic_disk_convert_to_pos.geo_unrolled'
Info    : Stopped on Mon Nov 29 14:05:54 2021

      Error   : Unknown number option 'General.NativeFileChooser'
      Warning : Unable to open file 'Analytic_disk/Analytic_disk/analytic_disk.msh'
      Error   : 'Analytic_disk/analytic_disk_convert_to_pos.geo', line 2 : Unknown view 0
```
