### 2021-11-29 #3905 @CFGrote
Review started. Installation went ok with minor hickups (will post separate issue later). Running the first example gave me errors, see https://github.com/Magritte-code/Magritte/issues/54. I'd like to see this fixed before I continue.
