### 2021-11-29 6017cbf
Started implementing the basic collocation solver. TODO: think about file structure (boundary intensity)
