+++ src/model/lines/lines.hpp
@@ -24,8 +24,8 @@ struct Lines
 
 
 
-    Matrix<Real> emissivity;      ///< line emissivity    (p, lid)
-    Matrix<Real> opacity;         ///< line opacity       (p, lid)
+    Matrix<Real> emissivity;      ///< line emissivity    (p, lid) // emissivity integrated over line profile DIVIDED BY the frequency
+    Matrix<Real> opacity;         ///< line opacity       (p, lid) // opacity integrated over line profile    DIVIDED BY the frequency
     Matrix<Real> inverse_width;   ///< inverse line width (p, lid) //includes the factor 1/sqrt(2) for convenient use in gaussians
 
     void read  (const Io& io);
