### 2021-12-01 Magritte-code/Magritte#54 @FredDeCeuster
Dear @CFGrote, thanks for your issue!

Based on the output, I suspect that the origin of the issue is that the working directory `wdir` that you specified in the notebook is given by a relative path instead of an absolute path.

Would it be a sufficient solution if we add a comment on that in the example notebooks, or would you have a better solution?
