### 2021-12-01 Magritte-code/Magritte#54 @CFGrote
Hi @FredDeCeuster , yes I think adding some documentation to the notebook might be sufficient.

Thanks!
