### 2021-12-01 5d98664
First implementation of collocation solver. TODO: fix condition number by using the source function formulation
