+++ src/bindings/pybindings.cpp
@@ -119,6 +119,7 @@ PYBIND11_MODULE (core, module)
         .def ("compute_Jeff",                                                       &Model::compute_Jeff)
         .def ("compute_level_populations_from_stateq",                              &Model::compute_level_populations_from_stateq)
         .def ("compute_level_populations",                                          &Model::compute_level_populations)
+        .def ("compute_level_populations_collocation",                              &Model::compute_level_populations_collocation)
         .def ("compute_level_populations_multiresolution",                          &Model::compute_level_populations_multiresolution)
         .def ("compute_image",                                                      &Model::compute_image)
         .def ("set_eta_and_chi",                                                    &Model::set_eta_and_chi)
