### 2021-12-01 50b7744
Highlight necessity of absolute paths in docs (closes #54)
