### 2021-12-03 bf560ff
Still bugfixing the collocation method. Optically thick case seems to work. TODO: figure out how to change/work around the implicitly defined central difference
