### 2021-12-06 4f5d698
Got the simplest benchmark working; TODO: explicitly include doppler shift.
