+++ src/bindings/pybindings.cpp
@@ -116,6 +116,7 @@ PYBIND11_MODULE (core, module)
         // .def ("compute_radiation_field",                                            &Model::compute_radiation_field)
         .def ("compute_radiation_field_feautrier_order_2",                          &Model::compute_radiation_field_feautrier_order_2)
         .def ("compute_radiation_field_shortchar_order_0",                          &Model::compute_radiation_field_shortchar_order_0)
+        .def ("compute_radiation_field_collocation",                                &Model::compute_radiation_field_collocation)
         .def ("compute_Jeff",                                                       &Model::compute_Jeff)
         .def ("compute_level_populations_from_stateq",                              &Model::compute_level_populations_from_stateq)
         .def ("compute_level_populations",                                          &Model::compute_level_populations)
