+++ src/model/model.hpp
@@ -20,7 +20,7 @@ struct Model
 {
     const Size MIN_INTERPOLATION_POINTS=16;   ///< Minimum number of points used during interpolation (~=average nb neighors in voronoi grid)
 
-    const Size INTERPOLATION_POINTS_1D=6;     ///< Number of interpolation points to use for 1D computations (spherical symmetry)
+    const Size INTERPOLATION_POINTS_1D=8;     ///< Number of interpolation points to use for 1D computations (spherical symmetry)
     const Size INTERPOLATION_POINTS_3D=16;    ///< Number of interpolation points to use for 3D computations (general geometry)
 
 
@@ -100,6 +100,7 @@ struct Model
     int compute_radiation_field                   ();
     int compute_radiation_field_feautrier_order_2 ();
     int compute_radiation_field_shortchar_order_0 ();
+    int compute_radiation_field_collocation       ();
     int compute_Jeff                              ();
     int compute_level_populations_from_stateq     ();
     int compute_level_populations                 (
