### 2021-12-08 Magritte-code/Magritte#56 @CFGrote
JOSS Review (Documentation and Examples)
I'd add a bit more information to the README.md regarding the field(s) of research in which Magritte can be applied. Is it limited to the astrophysical domain or more broadly applicable (e.g. plasma physics).

## API docs
Rudimentary. Only very few cpp and no py objects are documented.
- [x] ~Add API documentation for the classes and functions exposed to the user.~
**Edit**: Same comment as made by @gonsie over at https://github.com/openjournals/joss-reviews/issues/3905#issuecomment-996978642.

- [x] Add guidelines for contributors (e.g. point to issue labels, recomment on forking vs. becoming a  repo member etc)

# Examples
## docs/src/1_examples/0_creating_models/0_create_analytic_disk.ipynb
- [x] Typo in cell [1]: "We create a Magritte model ~form~ from..."
- [x] Function `turbulence(rr)`: `rr` is not used in the function body
- [x] wdir issue: I modified cell [3] to
```
wdir = os.path.abspath(os.path.join(os.getcwd(), 'Analytic_disk'))
```
Cell [6] then produces filenames such as "Analytic_diskco.txt" (without the "/" dir separator). I suggest to make use of `os.path.join()` to
construct paths (here and everywhere else in the code if applicable).

- [x] Error in cell [15]:
```
%%capture
mesher.convert_msh_to_pos (f'{bmesh_name}.msh', replace=True)
Error   : Unknown number option 'General.NativeFileChooser'
Error   : Unknown entity 1 of dimension 0
Error   : Could not read vertices
Error   : Error loading '/home/grotec/Repositories/JOSSReviews/Magritte/docs/src/1_examples/0_creating_models/Analytic_disk/analytic_disk.msh'
Error   : '/home/grotec/Repositories/JOSSReviews/Magritte/docs/src/1_examples/0_creating_models/Analytic_disk/analytic_disk_convert_to_pos.geo', line 2 : Unknown view 0
```

## 1_create_analytic_spiral.ipynb
- [x] Same typo as above
- [x] Same path issue as above
- [x] Error:
```
mesher.convert_msh_to_pos (f'{bmesh_name}.msh', replace=True)
Error   : Unknown number option 'General.NativeFileChooser'
Error   : Unknown entity 1 of dimension 0
Error   : Could not read vertices
Error   : Error loading '/home/grotec/Repositories/JOSSReviews/Magritte/docs/src/1_examples/0_creating_models/Analytic_spiral/analytic_spiral.msh'
Error   : '/home/grotec/Repositories/JOSSReviews/Magritte/docs/src/1_examples/0_creating_models/Analytic_spiral/analytic_spiral_convert_to_pos.geo', line 2 : Unknown view 0
```

I'll be happy to test the other examples but this should probably be sorted out first. Maybe worth a separate issue.
