### 2021-12-08 #3905 @CFGrote
Review comments on Documentation and Examples in https://github.com/Magritte-code/Magritte/issues/56
