### 2021-12-08 Magritte-code/Magritte#57 @CFGrote
JOSS Review (Software paper)
## Summary
A non-expert reader may find the summary too technical and dominated by domain specific jargon. I'd move the first three sentences of the "Statement of need" section to the "Summary" section to make the latter more accessible. Otherwise, the section (as the remainder of the paper) is extremely clearly written and transparent.
There is a typo on page 3, line 43: ~Furtherore~ Furthermore
As mentioned elsewhere, please delineate precisely the realm of applicability of the code, i.e. if it can be applied to objects other than
astrophysical objects.

## Statement of need
Figure 1 is referenced in the text but it remains unclear how to connect it to the statement "Their intricate morpho-kinematics, moreover, makes their appearance in observations far from evident (e.g., Figure 1)."

The authors claim that their software can be used to characterize complex structures. The inclined reader will find tests and benchmarks in the referenced work (in particular De Ceuster 2019 and 2020). This could be emphasized again in the software paper. Are these benchmarks obtained from other radiative transport codes or did you also compare to other methods such as e.g. particle-in-cell simulations?

## Future work
Out of personal curiosity, is there a principle difficulty in modifying the method to simulate **continuum** radiative transfer? Would this have to be considered in the mentioned treatment of non-linear coupling?
