### 2021-12-08 #3905 @CFGrote
Software paper review in https://github.com/Magritte-code/Magritte/issues/57
