### 2021-12-09 8ca4bf7
Found issue with boundary condition defined on opposite side.
