### 2021-12-09 794338d
Adding slope to rbf seems to alleviate numerical issues. TODO: define sensible slope for spherically symmetric models.
