### 2021-12-14 22d0324
First attempt at implementing equations in comoving frame. TODO: figure out what to doppler shift exactly
