+++ joss_paper_1.0/paper.md
@@ -15,32 +15,38 @@ authors:
   - name: Atulit Srivastava
     #orcid: 0000-0003-3997-9642
     affiliation: "2"
+  - name: Ward Homan
+    orcid: 0000-0001-7314-5081
+    affiliation: "3"
+  - name: Jan Bolte
+    orcid: 0000-0002-7991-9663
+    affiliation: "2"
   - name: Jeremy Yates
     orcid: 0000-0003-1954-8749
-    affiliation: "2"
+    affiliation: "1"
   - name: Leen Decin
     orcid: 0000-0002-5342-8612
-    affiliation: "2, 3"
+    affiliation: "2, 4"
   - name: Peter Boyle
     orcid: 0000-0002-8960-1587
-    affiliation: "4, 5"
+    affiliation: "5, 6"
   - name: James Hetherington
     orcid: 0000-0001-6993-0319
-    affiliation: "6, 7"
+    affiliation: "7"
 affiliations:
  - name: Department of Physics and Astronomy, University College London, London, UK
    index: 1
  - name: Institute of Astronomy, KU Leuven, Leuven, Belgium
    index: 2
- - name: School of Chemistry, University of Leeds, Leeds, UK
+ - name: Institut d’Astronomie et d’Astrophysique, Université Libre de Bruxelles, Brussels, Belgium
    index: 3
- - name: School of Physics and Astronomy, The University of Edinburgh, Edinburgh, UK
+ - name: School of Chemistry, University of Leeds, Leeds, UK
    index: 4
- - name: Brookhaven National Laboratory, NY, USA
+ - name: School of Physics and Astronomy, The University of Edinburgh, Edinburgh, UK
    index: 5
- - name: Department of Computer Science, University College London, London, UK
+ - name: Brookhaven National Laboratory, NY, USA
    index: 6
- - name: The Alan Turing Institute, London, UK
+ - name: Department of Computer Science, University College London, London, UK
    index: 7
 date: 1 October 2021
 bibliography: paper.bib
@@ -55,8 +61,13 @@ bibliography: paper.bib
 
 # Summary
 
-Magritte is a software library for 3D radiative transfer and synthetic observations,
-which currently focusses on atomic and molecular line transfer.
+Electromagnetic radiation is a key component in many astrophysical simulations.
+Not only does it dictate what we can or cannot observe, it can provide radiation
+pressure, efficient heating and cooling mechanisms, and opens up a range of new
+chemical pathways due to photo-reactions. Magritte is a software library that can be
+used as a general-purpose radiative transfer solver, but was particularly designed for
+line radiative transfer in complex 3D morphologies, such as, for instance,
+encountered in the stellar winds around evolved stars [see @Decin:2021].
 It is mainly written in C++ and can either be used as a Python package or
 as a C++ library. To compute the radiation field, a deterministic ray-tracer
 and a formal solver are employed, i.e. rays are traced through the model and the
@@ -71,23 +82,17 @@ Futhermore, tools are provided to optimise different input meshes for radiative 
 
 # Statement of need
 
-Electromagnetic radiation is a key component in many astrophysical simulations.
-Not only does it dictate what we can or cannot observe, it can provide radiation
-pressure, efficient heating and cooling mechanisms, and opens up a range of new
-chemical pathways due to photo-reactions. Magritte can be used as a
-general-purpose radiative transfer solver, but was particularly designed for
-line radiative transfer in complex 3D morphologies, such as, for instance,
-encountered in the stellar winds around evolved stars [see @Decin:2021].
 Recent high-resolution observations exposed the intricate and intrinsically 3D
 morphologies of these objects [@Decin:2020]. The sheer amount of complexity that is
 observed, makes it difficult to interpret the observations and necessitates the use of
[…]
