### 2021-12-14 751b2c9
Added comment on continuum.
