### 2021-12-14 Magritte-code/Magritte#57 @FredDeCeuster
Thanks, I've tried to address the issues in the latest commit.

Regarding your question: indeed also continuum radiation can relatively easily be accounted for (and will be added soon).
