### 2021-12-16 77c9bf0
Improved examples following #56
