### 2021-12-16 bf61e71
Added info for contributors.
