+++ CONTRIBUTING.md
@@ -1,0 +1,35 @@
+# Contributing to Magritte
+---
+
+Thank you for considering to contribute to Magritte! We highly appreciate any efforts to improve our software.
+In case you have any suggestions, questions, possible bugs, or feature requests, please, open an issue [here](https://github.com/Magritte-code/Magritte/issues).
+
+
+## Contributing examples and use cases
+---
+If you have used Magritte in a way that is different from any of [our examples](https://magritte.readthedocs.io/en/stable/1_examples/index.html), we highly encourage you to contribute to the [examples]()https://magritte.readthedocs.io/en/stable/1_examples/index.html in the documentation.
+You can do this by preparing your example as a jupyter notebook and creating a pull request (see below). I you are unsure about your example or how to add it, please contact us and we can help you.
+
+
+## Contributing new functionality
+---
+We have a clear development strategy for Magritte, which is closely related to our reseach. Several features are being developed for several projects, all of which might not yet be publicly visible. Therefore, to prevent development overlap, we advise to contact us (e.g. by creating an [issue](https://github.com/Magritte-code/Magritte/issues)), before starting to work on Magritte yourself. We are happy to share our preliminary features if they could help you, or welcome you to our develoment team to colaboratively work on our projects.
+
+
+## Contributing bug fixes
+---
+For bug fixes or smaller contributions, we gladly accept pull requests (see below).
+
+
+## Creating pull requests
+---
+The advised procedure for creating pull requests for Magritte goes as follows:
+
+- Fork the Magritte repository and create a new branch for your work;
+- Add your changes and commit them.
+- Run the benchmarks and/or the relevant examples.
+- Open a pull request for your branch on Github.
+
+Also, have a look at [this guide](https://docs.github.com/en/get-started/quickstart/contributing-to-projects) on forking in Github.
+
+_Thank you for contributing to Magritte!_
