### 2021-12-16 ea7b803
Implemented frequency quadratures; also started properly implementing doppler shift. TODO: figure out why these terms are so large...
