### 2021-12-17 Magritte-code/Magritte#56 @CFGrote
>     * I suspect that the error might have something to do with the file names. Could you show me the content of the model directory `'/home/grotec/Repositories/JOSSReviews/Magritte/docs/src/1_examples/0_creating_models/Analytic_disk`?

See attached and

![image](https://user-images.githubusercontent.com/15476807/146559940-383be15f-8751-466e-a5cb-50530b2962b8.png)

. Thanks for looking into this.
[co.txt](https://github.com/Magritte-code/Magritte/files/7735747/co.txt)
