### 2021-12-17 Magritte-code/Magritte#56 @FredDeCeuster
Thanks!
So no mesh files seem to have been created.
Since I presume you also made alterations to the notebook, could you send me the entire notebook please?
