### 2021-12-17 Magritte-code/Magritte#56 @CFGrote
in my fork at [docs/src/1_examples/0_creating_models/0_create_analytic_disk.ipynb](https://github.com/mpievolbio-scicomp/Magritte/blob/stable/docs/src/1_examples/0_creating_models/0_create_analytic_disk.ipynb)
