### 2021-12-17 Magritte-code/Magritte#56 @FredDeCeuster
Interesting... That works for me with a fresh install...

Could you comment out the `%%capture` in the cells above please?
And in cell [15] change `replace=True` to `replace=False`?
That should give us more info as to what is happening...
Cell [14] should create a .msh file in the directory `wdir`.
This normally gets replaced in cell [15], unless you specify `replace=False`.
So now it should still be there. Could you also send me that file?

Perhaps relevant, how did you install `gmsh`?

Many thanks!
