### 2021-12-17 #3905 @gonsie
I find the documentation of the code base somewhat lacking. There is almost no API documentation. While the examples do a decent job of walking someone through creating a model, the text itself is limited and doesn't explain much. A note in the documentation points to the code base... but there are very few comments in the code.
