+++ magritte/mesher.py
@@ -22,21 +22,38 @@ def relocate_indices(arr, p):
 class Mesh:
     """
     Magritte custom mesh class.
+    
+    Methods
+    -------
+    get_edges(self)
+        Returns a list with all the edges as index pairs of the points the edges connect.
     """
     def __init__(self, meshFile):
+        """
+        Read the meshFile.
+        
+        Parameters
+        ----------
+        meshFile : str
+            File name of the mesh file.
+        """
         self.mesh      = meshio.read(meshFile)
         self.points    = self.mesh.points
         self.tetras    = self.mesh.cells_dict['tetra']
         self.edges     = self.get_edges()
         self.neighbors = self.get_neighbors()
         self.boundary  = self.get_boundary()
+        
         # Remove non-connected points
         non_point = self.get_non_point()
         while not (non_point is None):
             print(non_point)
             self.del_non_point(non_point)
             non_point = self.get_non_point()
+            
+        return
 
+    
     def get_edges(self):
         edges = set([])
         for tetra in self.tetras:
@@ -121,10 +138,15 @@ class Mesh:
 def run(command):
     """
     Run command in shell and continuously print its output.
-    :param command: the command to run in the shell.
+    
+    Parameters
+    ----------
+    command : str
+        The command to run in the shell.
     """
     # Run command pipe output
     process = Popen(command, stdout=PIPE, shell=True)
+    
     # Continuously read output and print
     while True:
         # Extract output
@@ -134,44 +156,74 @@ def run(command):
             break
         else:
             print(line.decode("utf-8"))
+    
+    return
 
 
-def convert_msh_to_pos(meshName, replace: bool = False):
+def convert_msh_to_pos(meshName, replace:bool=False):
     """
-    Convert .msh to .pos background  mesh file.
-    :param modelName: path to the .msh file to convert.
-    :param replace: remove the original .msh file if True.
+    Convert a .msh file to a .pos file.
+    
+    Parameters
+    ----------
+    modelName : str
+        Path to the .msh file to convert.
+    replace : bool
+        Whether or not to remove (and thus replace) the original .msh file.
     """
     # Remove extension from meshName
     meshName, extension = os.path.splitext(meshName)
+    
     # create the converision gmsh script file
     conversion_script = f'{meshName}_convert_to_pos.geo'
+    
     # Get the path to this folder
     thisFolder = os.path.dirname(os.path.abspath(__file__))
+    
+    # Extract the template
     with open(f'{thisFolder}/templates/convert_to_pos.template', 'r') as file:
         template = Template(file.read())
+    
+    # Fill out the template and write the conversion script
     with open(conversion_script,                                 'w') as file:
         file.write(template.substitute(FILE_NAME=meshName))
-    # run gmsh in a subprocess to convert the background mesh to the .pos format
+    
+    # Run gmsh in a subprocess to convert the background mesh to the .pos format
     run(f'gmsh -0 {conversion_script}')
+    
     # Remove the auxiliary file that is created (geo_unrolled) and the script file
     os.remove(f'{meshName}_convert_to_pos.geo_unrolled')
     os.remove(conversion_script)
     if replace:
         os.remove(f"{meshName}.msh")
+        
+    return
 
 
 def boundary_cuboid (minVec, maxVec):
     """
-    Create the gmsh script for a cuboid element.
-    :param minVec: lower cuboid vector.
-    :param maxVec: upper cuboid vector.
-    :return: gmsh script string.
+    Retuns the gmsh script for a cuboid element that can be used as boundary.
+    
+    Parameters
+    ----------
[…]
