+++ docs/src/3_python_api_documentation/magritte.rst
@@ -1,0 +1,19 @@
+magritte package
+================
+
+.. automodule:: magritte
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+.. toctree::
+   :maxdepth: 0
+
+   magritte.core
+   magritte.mesher
+   magritte.plot
+   magritte.setup
+   magritte.tools
