+++ docs/src/3_python_api_documentation/magritte.mesher.rst
@@ -1,0 +1,7 @@
+magritte.mesher module
+======================
+
+.. automodule:: magritte.mesher
+   :members:
+   :undoc-members:
+   :show-inheritance:

[…]
