### 2021-12-17 aa81975
Added Python API documentation (fix RTD).
