### 2021-12-17 39659db
Added Python API documentation (fix RTD 2).
