### 2021-12-20 b7b1693
Fixed numerical instability by multiplying the boundary conditions by a factor.
