+++ docs/src/conf.py
@@ -13,7 +13,6 @@
 import os
 import sys
 sys.path.insert(0, os.path.abspath('.'))
-sys.path.insert(0, os.path.abspath('../..'))
 import datetime
 year = datetime.datetime.now().year
 
@@ -70,9 +69,6 @@ html_theme_options = {'logo_only': True}
 html_static_path = ['_static']
 
 
-
-
-
 # -- Run doxygen on READTHEDOCS server ---------------------------------------
 
 from subprocess import call
@@ -84,36 +80,8 @@ autodoc_member_order = "bysource"
 autodoc_default_flags = ["members"]
 autosummary_generate = True
 
-read_the_docs_build = os.environ.get('READTHEDOCS', None) == 'True'
+# read_the_docs_build = os.environ.get('READTHEDOCS', None) == 'True'
+# if read_the_docs_build:
 
-# if not read_the_docs_build:
-    
-def run_apidoc(_):
-    try:
-        from sphinx.ext.apidoc import main
-    except ImportError:
-        from sphinx.apidoc     import main
-
-    cur_dir = os.path.abspath(os.path.dirname(__file__))
-
-    api_doc_dir = os.path.join(cur_dir, "3_python_api_documentation")
-    module = os.path.join(cur_dir, "../..", "magritte")
-    ignore = []
-
-    os.environ["SPHINX_APIDOC_OPTIONS"] = "members, show-inheritance"
-    
-    # (Temporarily) make a symlink to core.so in the magritte module directory.
-    os.symlink(
-        os.path.join(cur_dir, "../..",      "bin/core.so"),
-        os.path.join(cur_dir, "../..", "magritte/core.so")
-    )
-    
-    main(["-M", "-f", "-e", "-T", "-d 0", "-o", api_doc_dir, module, *ignore])
-
-    # Remove the symlink to the core.so file
-    os.unlink(
-        os.path.join(cur_dir, "../..", "magritte/core.so")
-    )
-        
-def setup(app):
-    app.connect("builder-inited", run_apidoc)
+# Build Magritte to allow sphinx to extract the documentation.
+call('bash ../../build.sh',     shell=True)
