+++ docs/src/conf.py
@@ -84,4 +84,5 @@ autosummary_generate = True
 # if read_the_docs_build:
 
 # Build Magritte to allow sphinx to extract the documentation.
+call('conda activate magritte', shell=True)
 call('bash ../../build.sh',     shell=True)
