+++ docs/src/conf.py
@@ -83,6 +83,6 @@ autosummary_generate = True
 # read_the_docs_build = os.environ.get('READTHEDOCS', None) == 'True'
 # if read_the_docs_build:
 
-# Build Magritte to allow sphinx to extract the documentation.
-call('conda env list',          shell=True)
-call('bash ../../build.sh',     shell=True)
+# # Build Magritte to allow sphinx to extract the documentation.
+# call('conda env list',          shell=True)
+# call('bash ../../build.sh',     shell=True)
