+++ .readthedocs.yaml
@@ -16,8 +16,16 @@ build:
   tools:
     python: "mambaforge-4.10"
 
+# Set conda environment
 conda:
   environment: dependencies/conda_env.yml
+
+# Install Magritte in conda environment
+python:
+  version: "3.7"
+  install:
+    - method: setuptools
+      path: .
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
