+++ .readthedocs.yaml
@@ -22,7 +22,6 @@ conda:
 
 # Install Magritte in conda environment
 python:
-  version: "3.7"
   install:
     - method: setuptools
       path: .
