+++ setup.py
@@ -8,12 +8,13 @@ from setuptools.command.build_ext import build_ext
 from subprocess import call
 
 this_dir = os.path.dirname(os.path.realpath(__file__))
-print(os.listdir(os.getcwd()))
+print('ls curdir', os.listdir(os.getcwd()))
+print('ls curdir/src', os.listdir(f'{os.getcwd()}/src'))
 
 with open(os.path.join(this_dir, "./README.md"), "r") as file:
     long_description = file.read()
 
-with open("./src/configure.hpp", "r") as file:
+with open("src/configure.hpp", "r") as file:
     for line in file.readlines():
         if 'MAGRITTE_VERSION' in line:
             # Get the version, which is between quotes ("")
@@ -27,9 +28,9 @@ class my_build_ext(build_ext):
         Build magritte C++ core module.
         '''
         # Compile module
-        call('bash ./compile.sh', shell=True)
+        call('bash compile.sh', shell=True)
         # Copy the already-compiled core.so file.
-        shutil.copyfile('./bin/core.so', self.get_ext_fullpath(ext.name))
+        shutil.copyfile('bin/core.so', self.get_ext_fullpath(ext.name))
 
 
 setup(
