### 2022-01-05 b535929
Fixed error in frequency derivative.
