### 2022-01-05 Magritte-code/Magritte#56 @CFGrote
I tried again with `%%capture` magics removed. This time the notebook runs through to the end. Not sure what caused the earlier errors.

I installed gmsh via apt package manager (debian 10.2).
