### 2022-01-05 Magritte-code/Magritte#56 @CFGrote
> Dear @CFGrote, many thanks for your comments! I have tried to address most of these issues in the latest commit. However, we are currently still working on the API documentation.
>
> Some further comments:
>
>     * The `turbulence` function does indeed not use `rr`. However, since it is in general a function of `rr`, it seemed appropriate to include it here. That way, only the function body needs to be altered to change the turbulence, and not also cell [16] for instance.
>
>     * I prefer not to use `os.getcwd()` to keep the model files out of the source code. (The `os.path.join` suggestion was very helpful, thanks!).
>
>
> And a question:
>
>     * I suspect that the error might have something to do with the file names. Could you show me the content of the model directory `'/home/grotec/Repositories/JOSSReviews/Magritte/docs/src/1_examples/0_creating_models/Analytic_disk`?

ok, sounds reasonable to me. thanks for clarifying!
