### 2022-01-05 #3905 @CFGrote
@FredDeCeuster , can you address the pdf compile error?

> PDF failed to compile for issue #3905 with the following error:
>
> ```
>  Cloning into '0f2c2d2aa683fed30ea68102'...
> ```

@FredDeCeuster, can you address this?
