### 2022-01-05 b623aff
Applied inverse of diagonal as left preconditioner for the matrix.
