### 2022-01-07 f50b489
Fixed some bugs in the observer frame formulation. TODO: figure out how to choose the quadratures.
