+++ src/solver/collocation/collocation.hpp
@@ -11,14 +11,25 @@ class Collocation
 {
 
 private:
-    const bool USING_COMOVING_FRAME=true;//toggle for using the comoving frame
+    // const bool USING_COMOVING_FRAME=true;//toggle for using the comoving frame
+    const bool USING_COMOVING_FRAME=false;// exponent range for rhs is a bit problematic
 
     //internal parameters for the basis functions
     const Real TRUNCATION_SIGMA=5.0; //< if Delta(freq)>trunc_sigma*freq_width just set the result from the gaussian basis function to zero (as it is almost zero)
     const Real SLOPE_FACTOR=1.0;//1.0;//For stabilizing the equations a bit (making them somewhat more diagonally dominant);
-    const Real SLOPE_STEEPNESS=4.0;//slope steepness, also for stabilizing the equations; will practically be divided by 4 at x=0.
-    const Real FREQ_OFFSET=-0.2;//offset for the frequency basis functions (in terms of the inverse line width (mind the sqrt(2)))-in comoving frame
+    const Real SLOPE_STEEPNESS=4.0;//4.0;//slope steepness, also for stabilizing the equations; will practically be divided by 4 at x=0.
+    const Real FREQ_OFFSET=0.0;//-0.2;//offset for the frequency basis functions (in terms of the inverse line width (mind the sqrt(2)))-in comoving frame
     //FIXME: freq offset makes the freq cutoff condition a bit nonsymmetrical
+    const Real SLOPE_THRESHOLD=7.0*2.0;//The threshold for determining whether we need to add a slope to the radial basis function
+    //should be related to the max derivative of the rbf, the number of neighbors (for diagonal dominance reasons). Note: this comes from a global bound
+    //if opacity*typical radius>SLOPE_THRESHOLD, then do not add the slope.
+    // const Real QUADRATURE_MIN=-0.5;
+    // const Real QUADRATURE_MAX=0.5;
+    // TODO: define quadratures such that we can probe the entire space... abs max doppler shift-1 should not exceed the quadrature bounds
+
+    vector<vector<vector<bool>>> rbf_using_slope; //For every ray, for every point, for every frequency, denotes whether radial basis functions at a given point need to include the slope
+    //currently using some global bound; MAYBE TODO: use the exact formulation for each basis function; this will need some code duplication (explicit rbf derivative without slope), and will obviously also depend on the exact direction (hatn dot grad)
+    //frequency dependence is due to multiplying with the opacity and angular dependence might come into play when the rbf basis function could depend on the angle
 
     // DEPRECATED: makes that the basis functions are too peaked for my liking
     // const Real DISTANCE_EXPONENT=1.0/5.0; //Exponent for redistributing the distances
@@ -83,15 +94,18 @@ public:
 
     inline void setup_basis(Model& model);
 
+
+    inline Real get_slope(Real x, Real rayidx, Real pointidx, Real freqidx);
+    inline Real get_slope_derivative(Real x, Real rayidx, Real pointidx, Real freqidx);
 // The complete radial basis function at an index (dir, nu, pid) is a product of the ones defined below
 // basis functions for dir, nu, pid(implicitly x,y,z)
     inline Real basis_direction(Size rayindex);
     inline Real basis_freq(Size rayidx, Size freqidx, Size pointidx, Real currfreq);
-    inline Real basis_point(Size pointid, Vector3D& location, Size rayindex, Geometry& geometry);
+    inline Real basis_point(Size pointid, Vector3D& location, Size rayindex, Size freqidx, Geometry& geometry);
 
     // basis function derivatives for nu, position
[…]
