### 2022-01-12 51e61e0
First attempt at introducing a new frequency quadrature.
