### 2022-01-12 Magritte-code/Magritte#57 @CFGrote
ok, paper looks good to me now.
