### 2022-01-12 #3905 @CFGrote
@FredDeCeuster, just to make sure I'm not missing anything, are you still working on the API documentation or do you see it as ready for a (re-)review?
