### 2022-01-26 55879f3
Separated the matrix index into row and column parts. Also some prep for 2nd order feautrier.
