+++ src/bindings/pybindings.cpp
@@ -26,12 +26,25 @@ PYBIND11_MODULE (core, module)
     // Module docstring
     module.doc() = "Core module of Magritte: a modern software library for 3D radiative transfer.";
 
-    module.def(    "n_threads_avail", &paracabs::multi_threading::    n_threads_avail);
-    module.def("set_n_threads_avail", &paracabs::multi_threading::set_n_threads_avail);
+
+    module.def(
+        "n_threads_avail",
+        &paracabs::multi_threading::n_threads_avail,
+        "Get the number of available threads (using OpenMP)."
+    );
+
+
+    module.def(
+        "set_n_threads_avail",
+        &paracabs::multi_threading::set_n_threads_avail,
+        "Set the number of available threads (using OpenMP)."
+    );
+
 
     // Define vector types
     py::bind_vector<vector<LineProducingSpecies>> (module, "vLineProducingSpecies");
     py::bind_vector<vector<CollisionPartner>>     (module, "vCollisionPartner");
+
 
     // Constants
     module.attr("CC")    = CC;
@@ -40,151 +53,229 @@ PYBIND11_MODULE (core, module)
     module.attr("AMU")   = AMU;
     module.attr("T_CMB") = T_CMB;
 
+
     // Io, base class
-    py::class_<Io> (module, "Io");
+    py::class_<Io> (module, "Io", "Abstract input/output base class.");
 
 
     // IoText
-    py::class_<IoText, Io> (module, "IoText")
+    py::class_<IoText, Io> (module, "IoText", "Intput/output class for text-based io, implemented in C++.")
         // attributes
-        .def_readonly ("io_file", &IoText::io_file)
+        .def_readonly (
+            "io_file",
+            &IoText::io_file,
+            "Name of the folder or file to read from or write to."
+        )
         // constructor
         .def (py::init<const string &>());
 
 
     #if (PYTHON_IO)
         // IoPython
-        py::class_<IoPython, Io> (module, "IoPython")
+        py::class_<IoPython, Io> (module, "IoPython", "Input/output class for io, implemented in Python.")
             // attributes
-            .def_readonly ("implementation", &IoPython::implementation)
-            .def_readonly ("io_file",        &IoPython::io_file)
+            .def_readonly (
+                "implementation",
+                &IoPython::implementation,
+                "Type of input/output. Either \"hdf5\" or \"text\"."
+            )
+            .def_readonly (
+                "io_file",
+                &IoPython::io_file,
+                "Name of the folder or file to read from or write to."
+            )
             // constructor
             .def (py::init<const string &, const string &>());
     #endif
 
+
     // Solver
     py::class_<Solver> (module, "Solver")
-        // attributes
-        // functions
-        // .def ("trace", &Solver::trace)
         // constructor
-        // .def (py::init<const Size&, const Size&, const Size&>());
         .def (py::init<>());
 
+
     // Image
-    py::class_<Image> (module, "Image")
+    py::class_<Image> (module, "Image", "Image class, 2D point cloud of intensities for each frequency bin.")
         // attributes
-        // functions
-        .def_readonly  ("ray_nr", &Image::ray_nr)
-        .def_readonly  ("ImX",    &Image::ImX)
-        .def_readonly  ("ImY",    &Image::ImY)
-        .def_readonly  ("I",      &Image::I)
+        .def_readonly  ("ray_nr", &Image::ray_nr, "Number of the ray along which the image is taken.")
+        .def_readonly  ("ImX",    &Image::ImX,    "X-coordinates of the points in the image plane.")
+        .def_readonly  ("ImY",    &Image::ImY,    "Y-coordinates of the points in the image plane.")
[…]
