+++ docs/src/3_python_api_documentation/magritte.mesher.rst
@@ -3,5 +3,7 @@ magritte.mesher module
 
 .. automodule:: magritte.mesher
    :members:
-   :undoc-members:
+   :exclude-members: get_pfs, get_nbs, norm, get_Gs, get_Ls, get_weights
+   :special-members: __init__
    :show-inheritance:
+
