+++ docs/src/3_python_api_documentation/magritte.core.rst
@@ -3,5 +3,7 @@ magritte.core module
 
 .. automodule:: magritte.core
    :members:
-   :undoc-members:
+   :exclude-members: BoundaryCondition, Solver, Lambda, MReal, TReal, VReal, VSize, VVector3D, vCollisionPartner, vLineProducingSpecies
    :show-inheritance:
+..   :undoc-members:
+
