+++ meta.yaml
@@ -1,0 +1,44 @@
+package:
+  name: "magritte"
+  version: "0.0.1"
+
+# build:
+  # noarch: python
+
+source:
+  git_url: https://github.com/Magritte-code/Magritte
+
+requirements:
+  build:
+    - gcc
+    - git
+    - cmake>=3.18
+    - mpich
+    - compilers
+  host:
+    - pip
+    - python
+  run:
+    - h5py
+    - numpy
+    - scipy
+    - astropy
+    - numba
+    - conda-forge:meshio
+    - conda-forge:healpy
+    - yt
+    - mpi4py
+    - tqdm
+    - ipywidgets
+    - palettable
+    - plotly
+
+test:
+  imports:
+    - magritte
+
+about:
+  home: https://github.com/Magritte-code/Magritte
+  license: GPL-3.0
+  license_family: GPL
+  summary: A modern software library for simulating radiation transport.
