+++ magritte/tools.py
@@ -4,6 +4,7 @@ import os
 from datetime          import datetime
 from time              import perf_counter
 from astropy.io        import fits
+from astropy           import units, constants
 from scipy.interpolate import griddata
 
 
@@ -261,6 +262,9 @@ def save_fits(
         npix_x     = 300,
         npix_y     = 300,
         method     = 'nearest',
+        dpc        = 1.0,
+        coord      = None,
+        f_rest     = 0.0
     ):
     """
     Save channel maps of synthetic observation (image) as a fits file.
@@ -279,6 +283,12 @@ def save_fits(
         Number of pixels in the image in the vertical (y) direction.
     method : str
         Method to interpolate the scattered intensity data onto a regular image grid.
+    dpc : float
+        Distance of source in parsec.
+    coord : str
+        Image centre coordinates. 
+    f_rest : float
+        Rest frequency of the transition.
     
     Returns
     -------
@@ -322,17 +332,25 @@ def save_fits(
     
     # Extract the spectral / velocity data
     freqs = np.array(model.radiation.frequencies.nu)[0]
-    f_ij  = np.mean(freqs)
+    f_cen = np.mean(freqs)
+    
+    # If no rest frequency is given,
+    # default to cetral frequency in the image.
+    if f_rest == 0.0:
+        f_rest = f_cen
+    
+    velos = (freqs - f_rest) / f_rest * constants.c.si.value
+    v_cen = (f_cen - f_rest) / f_rest * constants.c.si.value
 
     dpix_x = np.mean(np.diff(xs))
     dpix_y = np.mean(np.diff(ys))
-    dfreqs = np.diff(freqs)
+    dvelos = np.diff(velos)
     
-    if (np.abs(relative_error(np.max(dfreqs), np.min(dfreqs))) > 1.0e-9):
+    if (np.abs(relative_error(np.max(dvelos), np.min(dvelos))) > 1.0e-9):
         print('WARNING: No regularly spaced frequency bins!')
-        dfreq = None
+        dvelo = None
     else:
-        dfreq = np.mean(dfreqs)
+        dvelo = np.mean(dvelos)
     
     # Interpolate the scattered data to an image (regular grid)
     zs = np.zeros((nfreqs, npix_x, npix_y))
@@ -340,17 +358,89 @@ def save_fits(
         # Nearest neighbor interpolate scattered image data
         zs[f] = griddata((imx, imy), imI[:,f], (xs[None,:], ys[:,None]), method=method)
     
-    hdr = fits.Header()
-    hdr['NAXIS'   ] = 3        # dims of the cube
-    hdr['NAXIS1'  ] = nfreqs   # number of pixels along velocity-axis
-    hdr['NAXIS2'  ] = npix_x   # number of pixels along x axis (hor)
-    hdr['NAXIS3'  ] = npix_y   # number of pixels along y-axis (ver)
-    hdr['CENTFREQ'] = f_ij     # central frequency of the image [Hz]
-    hdr['DPIX_X'  ] = dpix_x   # pixel size in meter [m] along x-axis
-    hdr['DPIX_Y'  ] = dpix_y   # pixel size in meter [m] along y-axis
-    hdr['DFREQ'   ] = dfreq    # pixel size in Hertz [Hz] along frequency axis
-    hdr['IM_UNIT' ] = 'W m^-2 Hz^-1 ster^-1'
+    # Convert intensity from J/s/m/m/ster to Jy/pixel
+    zs = zs * dpix_x * dpix_y / ((dpc * units.parsec).si.value)**2 / 1.0e-26    
+            
+    if coord is None:
[…]
