### 2022-02-04 6257c93
First (bugged) implementation of the feautrier collocation solver. TODO: look at second derivative
