### 2022-02-17 #3905 @CFGrote
For me this looks good to go! My issues in https://github.com/Magritte-code/Magritte are resolved and I ticked all boxes above.
