### 2022-02-18 06e4de8
Initial commit for pruning basis functions. Does not compile yet.
