+++ src/solver/collocation/collocation.hpp
@@ -33,7 +33,10 @@ private:
     // TODO: define quadratures such that we can probe the entire space... abs max doppler shift-1 should not exceed the quadrature bounds
     const Real MIN_OPACITY=1.0*std::pow(10,-28);//FIXME: use instead some adaptive measure based on rbf_der/(opacity*rbf). of diagonal element
 
-
+    vector<vector<Size>> remaining_points_after_pruning;
+    // Size n_freq_bases=1;//denotes the number of frequency bases used
+    // // not necessary, as the size of frequency_quadrature will tell us the same
+    // // = number of evaluation frequencies
 
     vector<vector<vector<bool>>> rbf_using_slope; //For every ray, for every point, for every frequency, denotes whether radial basis functions at a given point need to include the slope
     //currently using some global bound; MAYBE TODO: use the exact formulation for each basis function; this will need some code duplication (explicit rbf derivative without slope), and will obviously also depend on the exact direction (hatn dot grad)
@@ -44,11 +47,21 @@ private:
     // const Real DISTANCE_EPS=0.1;//TODO: should actually be chosen adaptively (depending on how close the closest point lies)
     // vector<Real> half_min_dist; //Half the distance for
 
+    //FIXME: actually compute the necessary value; eval of 0.7 in nearest frequency
+    const Real FREQ_QUADRATURE_WING_DISTANCE_AMP_FACTOR=1.5;//In order to sample the line wings sufficiently in the case of doppler shifts, we increase the distance between the frequencies on the line wings.
+    const Size FREQ_QUADRATURE_CENTER_N_ELEMENTS=5;//the number of quadrature elements which should be in the line center
+    //TODO?: let this depend on parameters.nquads()?
     const Real FREQ_QUADRATURE_EDGE=2.0;//In order to make sure to sample the boundary condition well for all points, a safety param is included (enlarges the bounds for the frequency quadrature respectively subtracting and adding this to the min and max doppler shifts)
+    // const Real FREQ_QUADRATURE_CENTER_SEPARATION_DISTANCE=(2.0*FREQ_QUADRATURE_EDGE)/(FREQ_QUADRATURE_CENTER_N_ELEMENTS-1);//The separation distance of the central frequency quadrature points (in terms of line width)
+
+    Size n_freq_bases=0;//denotes the total number of frequencies we evaluate (=size frequency_quadrature=size frequency_inverse_width_quadrature)
     // vector<Real> min_raydir_doppler_shift;//for every direction, contains the minimum doppler shift in that direction
     // vector<Real> max_raydir_doppler_shift;//for every direction, contains the maximum doppler shift in that direction
-    vector<Real> frequency_quadrature;//for every line frequency, contians the frequency quadrature
+    vector<Size> center_line_freq_indices;//for every line, stores the frequency index corresponding to the center of the line.
+    vector<Real> frequency_quadrature;//for every line frequency, contains the frequency quadrature
     vector<Real> frequency_inverse_width_quadrature;//for every line frequency, contains the typical inverse width
+    // vector<Real> frequency_quadrature_in_line_widths;//for every line frequency, contains the frequency quadrature in terms of line widths compare to the center line frequency
+    //not a good idea, might be far too complex and might ruin the exact influence region of boundary conditions...
 
     //TODO: adaptively define this such that diagonal remains somewhat ?constant...?
     //ideas: should drastically reduce ill-conditionedness of a matrix, so should (implicitly) depend on the number of points/distance of the closest point/...
[…]
