### 2022-02-22 3550c2e
Currently only works for the previous direct solver bases. TODO: think about least-squares preconditioner.
