### 2022-02-23 ba94d7b
Ignoring issues related to preconditioning of the matrix, the least-squares formulation works.
