### 2022-02-28 40ad187
Implemented square collocation method with pruning + some preliminary fixes to the new frequency quadrature. TODO: more fixes
