### 2022-02-28 Magritte-code/Magritte#52 @gonsie
closed by commit bf61e71b3ff03e2fa56fe8c1598e403de53558c0
