### 2022-02-28 Magritte-code/Magritte#61 @gonsie
Missing details on analytics benchmark
This page is currently empty.
https://magritte.readthedocs.io/en/stable/2_benchmarks/0_analytic_benchmarks.html

This is related to the review in openjournals/joss-reviews#3905
