### 2022-02-28 #3905 @gonsie
Hi all (@xuanxu @FredDeCeuster), I'm sorry for my delayed response.

I've updated my checkmarks. Due to the complexity of building this, I am unable to verify the functionality (I do not have easy access to the gcc compiler with a proper python/conda environment). While the paper does not reference performance numbers, they are present in the documentation and (again) I am unable to confirm them.

I've filed issue Magritte-code/Magritte#61 about including details about the analytical benchmark in the docs. I would like this documented since the automated tests run only a few benchmarks. While the `tests/run_all_tests.sh` script refers to 'unit tests' and 'integration tests' those do not exist.

There are a lot of details in the documentation that could be added to the paper to flush it out. Currently the paper consists of only a summary, statement of need, and future work. While more details of the current software capabilities would be nice, it  is not a requirement for a Joss paper.

Once issue # 61 is complete, I will accept this paper.

Thank you.
