### 2022-03-01 c29fec3
Added info about analytic benchmarks addressing #61
