### 2022-03-01 Magritte-code/Magritte#61 @FredDeCeuster
Dear @gonsie, thanks for your issue report!
I've added some more details on the analytic benchmark page.
Let me know if I can provide any further information.
