+++ joss_paper_1.0/paper.bib
@@ -76,6 +76,7 @@
     pages = {arXiv:2011.13472},
     archivePrefix = {arXiv},
     eprint = {2011.13472},
+    doi = {10.1146/annurev-astro-090120-033712},
     primaryClass = {astro-ph.SR},
 }
 
