### 2022-03-01 #3905 @FredDeCeuster
I've checked and added the doi for Decin (2020).
https://github.com/Magritte-code/Magritte/commit/30b5d08536350c997bb30526b094c2612442c277
