### 2022-03-03 00bc87b
Stashing the current attempt for the collocation solver.
