+++ tests/benchmarks/analytic/all_constant_single_ray.py
@@ -14,11 +14,11 @@ import magritte.core     as magritte
 
 
 dimension = 1
-npoints   = 50
+npoints   = 40
 nrays     = 2
 nspecs    = 5
 nlspecs   = 1
-nquads    = 1
+nquads    = 21
 
 nH2  = 1.0E+12                 # [m^-3]
 nTT  = 1.0E+03                 # [m^-3]
@@ -26,6 +26,7 @@ temp = 4.5E+00                 # [K]
 turb = 0.0E+00                 # [m/s]
 dx   = 1.0E+12                 # [m]
 dv   = 0.0E+00 / magritte.CC   # [fraction of speed of light]
+# dv   = 2.5E+01 / magritte.CC   # [fraction of speed of light]
 
 
 def create_model ():
@@ -91,16 +92,26 @@ def run_model (nosave=False):
     timer3.start()
     model.compute_radiation_field_shortchar_order_0 ()
     timer3.stop()
+    model.compute_Jeff();
     u_0s = np.array(model.radiation.u)
+    sum_J_0s=np.array(model.lines.lineProducingSpecies[0].Jlin)
 
     timer4 = tools.Timer('feautrier 2  ')
     timer4.start()
     model.compute_radiation_field_feautrier_order_2 ()
     timer4.stop()
+    model.compute_Jeff();
     u_2f = np.array(model.radiation.u)
+    J_2f = np.array(model.radiation.J)
+    sum_J_2f=np.array(model.lines.lineProducingSpecies[0].Jlin)
+    quads= np.array(model.lines.lineProducingSpecies[0].quadrature.weights)
+    # sum_J_2f= np.sum(J_2f*quads, axis=1)
+    #approx; because feautrier solver does not compute the intensities...
+    # I_2f = np.array(model.radiation.I)*u_2f[0,:,0]/u_0s[0,:,0]
 
     timer5 = tools.Timer('collocation  ')
     timer5.start()
+    # model.COMPUTING_SVD_AND_CONDITION_NUMBER=True
     model.compute_radiation_field_collocation ()
     timer5.stop()
     colf = np.array(model.lines.lineProducingSpecies[0].Jlin)
@@ -122,6 +133,38 @@ def run_model (nosave=False):
     chi = tools.lineOpacity    (ld, pop)[k] * phi
     src = tools.lineSource     (ld, pop)[k]
     bdy = tools.I_CMB          (frq)
+
+    velocities=np.array(model.geometry.points.velocity);
+    doppler_shifts=1+velocities[:,0];
+    doppler_shifts_dist=(doppler_shifts*frq-frq)/tools.dnu(ld, k, temp, (turb/magritte.CC)**2)
+    dopp_mult_factors=tools.profile(ld, k, temp, (turb/magritte.CC)**2, frq*doppler_shifts)/phi
+    # dopp_exp_factor=np.exp(-(1-dopp_mult_factors)**2)
+    # print("dopp exp factor", dopp_exp_factor)
+
+    # print(dopp_mult_factors)
+    # print(u_2f)
+    # rev_dopp_mult_factors=dopp_exp_factor[::-1]
+    #u is the sum of two intensities on the opposite directions, so we actually need to doppler shift them individually
+    #also do not forget the background intensity; TODO;: figure out better formula
+    # u_2f_dopp=(I_2f[0,:,0]*dopp_mult_factors+bdy*(1-dopp_mult_factors)+I_2f[1,:,0]*rev_dopp_mult_factors+bdy*(1-rev_dopp_mult_factors))/2
[…]
